Fix CarNotificationListener logic to immediately group notifications when applicable.
Previously, we did not always update the group key of alert entries when assigning a new value to RankingMap, which resulted in state inconsistencies.
We relied on NotificationListnerService's MSG_ON_RANKING_UPDATED event to invoke CarNotificationListener's onNotificationRankingUpdate method, which updates the group key of alert entries.
However, the timing of when NotificationListnerService's MSG_ON_RANKING_UPDATED is called is random, which caused the problem of taking a random amount of attempts of closing/reopening the shade panel to see the notifications grouped.
Now whenever a new value is assigned to RankingMap, we call onNotificationRankingUpdate and update the group key of alert entries at the same time.

Test: Manual
Bug: 138846306
Change-Id: Iad9e8fc8e8bc6ad337bb92250dba1ae9e9eb1a36