DO NOT MERGE Set message notification's title to the conversation title
A Notification using MessagingStyle should use the MessagingStyle's
conversationTitle attribute for its title. If the title is null, ~then~
it should fallback to using the last sender's name as the title.

This change will enable apps to post group message notifications since a
group conversation might have a name, or will at least list the members
of the group so the user doesn't confuse a group message with a 1-1
message.

Bug: 147767724
Test: manual
Change-Id: I1dc78321fb11f7c86d5bc73ef699393edfc98a3b