Fix shouldDismissOnSwipe logic error
The shouldDismissOnSwipe method was actually returning if the
notification should *not* be dismissed by swiping. Negated the logic so
the method returns a boolean consistent with its name -- if the
notification should be dismissed on swipe.

Also changed to using the Objects.equals() method to verify the
notification's category, since the category could be null.

Bug: 168533683
Test: manual - built on flame_car
Change-Id: I00a12c27ad0dc5df6d3c5f1da04779086494db7f