Order new notifications correctly
When the notification panel is visible we avoid regrouping or reordering
existing notifications. The current implementation tries to maintain a
ranking cache in order to do this, but mixing ranks from different
ranking maps leads to incorrect and inconsistent results. Instead we
should always use the up to date rank, and just insert new notifications
rather than doing a full sort.

Test: manual, atest PreprocessingManagerTest
Bug: 164493400
Change-Id: Idfb63ceec9db9a9afd0d647320c7bcc57647977c