Allow notifications actions button margin configuration
Add a dimen to add a right margin on the action view. This is so that notifications can be configured to avoid overlapping
with the dismiss button.

Unfortunately this doesn't work well to change dynamically based on
whether we're in rotary mode or not. OEMs should be able to configure
their notifications to avoid this kind of issue on their devices by
ensuring that the notification cards are wide enough.

Bug: 173728678
Test: manual
Change-Id: I8ad360474b02c62b9f56a522822daa9a328bbe80