Fix issue when removing a HUN based on an action.
The CarHeadsUpNotificationManager will automatically start a
process to create a notification in the Notification center when
the HUN was clicked. If the action is about removing the
notificaiton, the click listener process is still started and
creates the NCN.

This is a quick fix which stores the alertkeys to be removed
and intercepts the automatical process creating the NCN.

Updating the tests as maybeRemoveHeaudsUp needs to call getKey().

Bug: b/203784760
Test: manual (Seahawk) and run CarNotificationRoboTest

Change-Id: I37cb9656705b1b92e6710a446f25f5e079507d39