Add throttling and sorting of HUNs
A priority queue is used to show HUNs one after the other by using the OnHeadsUpNotificationStateChange listener. Priorities of HUN is assigned based on the AlertEntry's category and post time. While driving the HUNs in the queue might expire and be directly sent to Notification center if it exceds the time set by the OEM. CarHeadsUpNotificationQueueCallback is used to communicate changes to HUNs in the queue.

Bug: 236715713
Test: Manual, atest com.android.car.notification.CarHeadsUpNotificationQueueTest
Change-Id: I6eb0473722d4e2fd56b9d4ea97d93f7fd6ec80d2