Fix grouping new notifications when NC is open
Existing behavior was to add new notifications to Notification Group(NG)
with same group key(if already existed) when the Notification Center was
open. This created problem where the Notification was not visible since
the Adapter did not identify the NG as a grouped notification.

This fix will change the behavior of non summary notifications when
they are added and NC is open.
- If a NG is present with same group key and with children count
  greater than the minimum grouping threshold, the new notification
  will be added to that NG.
- If the minimum grouping threshold is not met, new Notification
  will be added as a new NG.
- If minimum grouping threshold is met and there is no existing
  group, new NG is created and all notifications are added to it and
  a summary notification is attached.

The new Notifications added when the NC is open will be shown in
recents and marked as seen.

Bug: 242553753
Test: Manual, added unit tests

Change-Id: Ib4eeab45bf634c3f18ea64e061f79b914c5797cb