Don't change the visibility of the RecyclerView in the NotificationShade
The GestureHandler to close the NotificationShade is attached to the
RecyclerView. When there are no notifications the RecyclerView was GONE
and the GestureHandler didn't receive any touch event and was not able
to hide the NotificationShade. This change fixes this.

Test: atest-src CarNotificationUnitTests, manually
Bug: 254060340
Change-Id: I233a1001842ba383f678015ac8ee9d3295ae58a1
Merged-In: I233a1001842ba383f678015ac8ee9d3295ae58a1