Fix updating seen notifications
Seen notifications that are updated when the notification panel is open
would create a new unseen notification leaving the seen notification
still in the view. This would create confusion as there would be 2
views representing the same notification.

This is not a problem if update happens when notification center is
closed since all notifications are regrouped before the notification
center is reopened.

The solution is to update additionalGroupAndRank method(called when a
change happens and the notification center is open), to remove the
notification from previous group, remove the group if no children are
present and add the update notification to a new group in unseen
section.

Test: manual - added video to bug, atest CarNotificationUnitTests
Bug: 295294064
Change-Id: Ib1342ef16cf294e3996fc24cd64bed7a38ee0e44