Fix crash and flaky failures in unit tests
2 classes failing: HeadsUpContainerViewTouchModeTest and
CarHeadsUpNotificationQueueTest

The crash in HeadsUpContainerViewTouchModeTest would crash the
instrumentation failing all consequent tests.
The fix was to remove using system services and Executors in
CarHeadsUpNotificationQueueTest and instead use mocks.

The flakiness in CarNotificationListenerTest was because
the tests started running before the listener was conncected.
The solution was to make the thread wait until the listener is
connected. Also noticed some tests expected for a method to be called
with null parameter since the new value is mapped to null, updated
this to use a variable instead to make the test more readable.

Bug: 299516091
Test: atest CarNotificationUnitTests --iterations 15
Change-Id: Ie2ae65bdf985b039b7cd4c8b0b27fd5ec395a76c