Fix memory leak in NotificationAdapter
OnConfigurationChange would create a new view and a new adapter, but the
adapter is registered in the PreprocessingManager and will never be
garbage collected. By unregistering the adapter from the recyclerview
and the PreprocessingManager this leak can be prevented.

Bug: 305930607
Test: Manual (memory dump)
Test: atest CarNotificationUnitTests
Change-Id: I4baf7ef2f9ecfc01de814a7a52ce49f1ef6838c4
(cherry picked from commit 96c2a7babd5a13e8acd19024fc5ca56d37826c33)