Fix removing HUN by sender on button click
When HUN is dismissed by clicking a button it is sent to the
notification center as a new notification. When notification
removes iteself on click, the remove call get's executed first
and the new notification is posted after(as a result of dismissing the
HUN). We track all dismissing HUNs and HUNs that are removed by the
sender in a concurrent HashSet and handle the cases seperately at the
time of dismiss.

Bug: 286785983
Test: atest CarNotificationUnitTests
Change-Id: I3f92327eb78dc66e677ae5caef643f0fb2672a4d