Allow only one CATEGORY_CALL to be shown as a HUN
Multiple calls coming in at the same time current
results in multiple overlapping HUNs and  only the
latest arrival call remaining interactible to user

- Allow Notification.CATEGORY_CALL notifications to be
dismissed when shown as a HUN.
- Show dismissed CATEGORY_CALL notifications in
NotificationView when its respective HUN is dismissed.
- Rename some method names and javadoc to be more
descriptive for their intended usage and effect (i.e,
maybeShowHun -> maybeShowOrScheduleHun to reflect
that it may be added to a throttled or pending queue.

Bug: 405200695
Test: atest CarNotificationUnitTests
Flag: EXEMPT bug fix only

Change-Id: I63d162ecf37e965df897b30f6f0bf03172a443c2